refactor(core): split agent_task god-file into module submodules
Splits src/core/agent_task.rs (1682 lines, 81 items — baselined GodFile)
into concern submodules under the existing agent_task/ dir (artifacts,
executor, matrix, outcome, policy, request, schema) with the test module
moved to tests.rs, each under structural thresholds. The agent_task.rs
module root is now a 54-line re-export aggregation so the large
crate::core::agent_task::* import surface is unchanged. Pure move;
baseline rows removed.
