+++ homeboy.json
@@ -735,7 +735,6 @@
         "structural::src/commands/report/failure_digest.rs::HighItemCount",
         "structural::src/commands/runs/gh_actions.rs::HighItemCount",
         "structural::src/commands::DirectorySprawl",
-        "structural::src/core/agent_task.rs::GodFile",
         "structural::src/core/agent_task_controller_service/artifacts.rs::HighItemCount",
         "structural::src/core/agent_task_controller_service/tests.rs::GodFile",
         "structural::src/core/agent_task_controller_service/tests.rs::HighItemCount",
