+++ src/core/agent_task/request.rs
@@ -1,0 +1,223 @@
+use serde::{Deserialize, Serialize};
+use serde_json::{Map, Value};
+
+use super::schema::request_schema;
+use super::{
+    AgentTaskArtifactDeclaration, AgentTaskExecutor, AgentTaskLimits, AgentTaskOutcome,
+    AgentTaskPolicy, AgentTaskSourceRef, AgentTaskWorkspace,
+};
+
+/// Provider capability payload used by extension discovery and durable run metadata.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct AgentTaskExecutorCapabilities {
+    pub backend: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub selector: Option<String>,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub capabilities: Vec<String>,
+    #[serde(default)]
+    pub supports_sync_completion: bool,
+    #[serde(default)]
+    pub supports_async_polling: bool,
+    #[serde(default)]
+    pub supports_streaming: bool,
+    #[serde(default)]
+    pub supports_cancel: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AgentTaskPreparedWorkspace {
+    pub mode: super::AgentTaskWorkspaceMode,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub root: Option<String>,
+    #[serde(default, skip_serializing_if = "Value::is_null")]
+    pub metadata: Value,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AgentTaskExecutionHandle {
+    #[serde(
+        default,
+        skip_serializing_if = "AgentTaskExecutionHandleKind::is_provider_run"
+    )]
+    pub kind: AgentTaskExecutionHandleKind,
+    pub task_id: String,
+    pub backend: String,
+    pub run_id: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub stream_uri: Option<String>,
+    #[serde(default, skip_serializing_if = "Value::is_null")]
+    pub metadata: Value,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Hash, Default)]
+#[serde(rename_all = "snake_case")]
+pub enum AgentTaskExecutionHandleKind {
+    QueuedRecord,
+    LocalPid,
+    RunnerJob,
+    #[default]
+    ProviderRun,
+}
+
+impl AgentTaskExecutionHandleKind {
+    pub(crate) fn is_provider_run(&self) -> bool {
+        matches!(self, Self::ProviderRun)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AgentTaskStart {
+    pub handle: AgentTaskExecutionHandle,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub outcome: Option<AgentTaskOutcome>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AgentTaskProgress {
+    pub handle: AgentTaskExecutionHandle,
+    pub state: AgentTaskExecutionState,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub events: Vec<AgentTaskProgressEvent>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub provider_payload: Option<Value>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub outcome: Option<AgentTaskOutcome>,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum AgentTaskExecutionState {
+    Queued,
+    Running,
+    Waiting,
+    Succeeded,
+    Failed,
+    Cancelled,
+}
+
+impl AgentTaskExecutionState {
+    pub fn is_terminal(self) -> bool {
+        matches!(self, Self::Succeeded | Self::Failed | Self::Cancelled)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AgentTaskProgressEvent {
+    pub kind: String,
+    pub message: String,
+    #[serde(default, skip_serializing_if = "Value::is_null")]
+    pub data: Value,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AgentTaskRequest {
+    #[serde(default = "request_schema")]
+    pub schema: String,
+    pub task_id: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub group_key: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub parent_plan_id: Option<String>,
+    pub executor: AgentTaskExecutor,
+    pub instructions: String,
+    #[serde(default, skip_serializing_if = "Value::is_null")]
+    pub inputs: Value,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub source_refs: Vec<AgentTaskSourceRef>,
+    #[serde(default)]
+    pub workspace: AgentTaskWorkspace,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub component_contracts: Vec<AgentTaskComponentContract>,
+    #[serde(default)]
+    pub policy: AgentTaskPolicy,
+    #[serde(default)]
+    pub limits: AgentTaskLimits,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub expected_artifacts: Vec<String>,
+    #[serde(
+        default,
+        alias = "artifactDeclarations",
+        skip_serializing_if = "Vec::is_empty"
+    )]
+    pub artifact_declarations: Vec<AgentTaskArtifactDeclaration>,
+    #[serde(default, skip_serializing_if = "Value::is_null")]
+    pub metadata: Value,
+}
+
+impl AgentTaskRequest {
+    pub fn canonical_artifact_declarations(&self) -> Vec<AgentTaskArtifactDeclaration> {
+        let mut declarations = Vec::new();
+        for declaration in &self.artifact_declarations {
+            if let Some(declaration) = declaration.canonical() {
+                push_artifact_declaration_once(&mut declarations, declaration);
+            }
+        }
+
+        for expected in &self.expected_artifacts {
+            if let Some(declaration) =
+                AgentTaskArtifactDeclaration::from_expected_artifact(expected)
+            {
+                push_artifact_declaration_once(&mut declarations, declaration);
+            }
+        }
+
+        declarations
+    }
+
+    pub fn normalize_artifact_declarations(&mut self) {
+        self.artifact_declarations = self.canonical_artifact_declarations();
+    }
+}
+
+fn push_artifact_declaration_once(
+    declarations: &mut Vec<AgentTaskArtifactDeclaration>,
+    declaration: AgentTaskArtifactDeclaration,
+) {
+    if declarations
+        .iter()
+        .any(|existing| existing.name == declaration.name)
+    {
+        return;
+    }
+    declarations.push(declaration);
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct AgentTaskComponentContract {
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub slug: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub path: Option<String>,
+    #[serde(
+        default,
+        rename = "loadAs",
+        alias = "load_as",
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub load_as: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub activate: Option<bool>,
+    #[serde(flatten)]
+    pub extra: Map<String, Value>,
+}
+
+#[cfg(test)]
+impl AgentTaskRequest {
+    pub(crate) fn redacted(&self) -> Self {
+        let policy = crate::core::redaction::RedactionPolicy::default();
+        let mut redacted = self.clone();
+        redacted.instructions = policy.redact_string(&redacted.instructions);
+        redacted.inputs = policy.redact_json(&redacted.inputs);
+        redacted.executor.config = policy.redact_json(&redacted.executor.config);
+        redacted.workspace.materialization =
+            policy.redact_json(&redacted.workspace.materialization);
+        redacted.artifact_declarations = redacted
+            .artifact_declarations
+            .into_iter()
+            .map(|declaration| declaration.redacted_with(&policy))
+            .collect();
+        redacted.metadata = policy.redact_json(&redacted.metadata);
+        redacted
+    }
+}
